ContentService/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-jose==3.3.0
pydantic==1.10.13
python-dotenv==1.0.1
httpx==0.27.0
redis==5.0.1
